[Fix] #65 Recyclerview Adapter분리
